Share ClickHouse row helpers across evm-dex and evm-nfts
