+++ evm-dex/src/sunpump.rs
@@ -1,15 +1,11 @@
+use common::clickhouse::{log_key, set_clock, set_template_log, set_template_tx};
 use common::{bytes_to_string, Encoding};
 use proto::pb::sunpump::v1::{self as sunpump, StorePool};
 use substreams::{pb::substreams::Clock, store::StoreGetProto};
 use substreams_database_change::tables::Tables;
 use substreams_ethereum::NULL_ADDRESS;
 
-use crate::{
-    logs::{log_key, set_template_log},
-    set_clock,
-    store::get_store_by_address,
-    transactions::set_template_tx,
-};
+use crate::store::get_store_by_address;
 
 // SunPump Processing
 pub fn process_events(encoding: &Encoding, tables: &mut Tables, clock: &Clock, events: &sunpump::Events, store: &StoreGetProto<StorePool>) {
